csharp: generate API resources into the Api package
