+++ csharp/Svix.Tests/SvixClientTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.Extensions.Logging.Abstractions;
+using Svix.Api;
 using Svix.Models;
 using Xunit;
