+++ csharp/Svix/Api/BackgroundTask.cs
@@ -3,7 +3,7 @@
 using Microsoft.Extensions.Logging;
 using Svix.Models;
 
-namespace Svix
+namespace Svix.Api
 {
     public class BackgroundTaskListOptions : SvixOptionsBase
     {
