+++ csharp/Svix/SvixClient.cs
@@ -1,5 +1,6 @@
 using System.Runtime.InteropServices;
 using Microsoft.Extensions.Logging;
+using Svix.Api;
 
 namespace Svix
 {
